+++ src/main/java/com/sb02/todoapp/model/ToDo.java
@@ -1,4 +1,27 @@
 package com.sb02.todoapp.model;
 
 public class ToDo {
+    private String name;
+    private String description;
+
+    public ToDo(String name, String description) {
+        this.name = name;
+        this.description = description;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
 }
